Actualizados repositorios y actualizado MongoReport para que cada reporte tenga una ID única.

+++ matrix-common/src/io/github/beelzebu/matrix/database/mongo/ReportDAO.java
@@ -11,7 +11,7 @@ import org.mongodb.morphia.query.Sort;
 /**
  * @author Beelzebu
  */
-public class ReportDAO extends BasicDAO<MongoReport, Integer> implements ReportManager {
+public class ReportDAO extends BasicDAO<MongoReport, Long> implements ReportManager {
 
     public ReportDAO(Datastore ds) {
         super(MongoReport.class, ds);
